Use sparse code reward

+++ nemo_aligner/models/nlp/gpt/reward_critic_clients.py
@@ -432,7 +432,7 @@ class CodeEvaluator:
 
         print([x for x in details])
         print(sum(details) / len(details))
-        return sum(details) / len(details)
+        return int(sum(details) / len(details))
 
     
     def task_mask(self, args, device):
